multiple path cfg

+++ V2Analyzer/interface/V2AnalyzerBase.h
@@ -172,7 +172,7 @@ private:
       double etaCutMax_;
       double etaHFLow_;
       double etaHFUpp_;
-      double order_;
+      int order_;
       int NTrkMin_;
       int NTrkMax_;
       bool doEffCorrection_;
